expression: set integer literals as int by default

+++ vlib/hash/fnv1a/fnv1a.v
@@ -1,8 +1,8 @@
 module fnv1a
 
 const (
-	fnv64_prime = 1099511628211
-	fnv64_offset_basis = 14695981039346656037
+	fnv64_prime = u64(1099511628211)
+	fnv64_offset_basis = u64(14695981039346656037)
 	fnv32_offset_basis = u32(2166136261)
 	fnv32_prime = u32(16777619)
 )
